LayoutTests/imported/w3c/web-platform-tests/css/css-nesting/cssom-expected.txt


FAIL Simple CSSOM manipulation of subrules ss.cssRules[0].insertRule is not a function. (In 'ss.cssRules[0].insertRule('& .b { color: green; }')', 'ss.cssRules[0].insertRule' is undefined)
FAIL Simple CSSOM manipulation of subrules 1 assert_throws_dom: function "() => { ss.cssRules[0].insertRule('& .broken {}', 3); }" threw object "TypeError: ss.cssRules[0].insertRule is not a function. (In 'ss.cssRules[0].insertRule('& .broken {}', 3)', 'ss.cssRules[0].insertRule' is undefined)" that is not a DOMException IndexSizeError: property "code" is equal to undefined, expected 1
FAIL Simple CSSOM manipulation of subrules 2 assert_throws_dom: function "() => { ss.cssRules[0].insertRule('& .broken {}', -1); }" threw object "TypeError: ss.cssRules[0].insertRule is not a function. (In 'ss.cssRules[0].insertRule('& .broken {}', -1)', 'ss.cssRules[0].insertRule' is undefined)" that is not a DOMException IndexSizeError: property "code" is equal to undefined, expected 1
FAIL Simple CSSOM manipulation of subrules 3 assert_throws_dom: function "() => { ss.cssRules[0].deleteRule(5); }" threw object "TypeError: ss.cssRules[0].deleteRule is not a function. (In 'ss.cssRules[0].deleteRule(5)', 'ss.cssRules[0].deleteRule' is undefined)" that is not a DOMException IndexSizeError: property "code" is equal to undefined, expected 1
FAIL Simple CSSOM manipulation of subrules 4 undefined is not an object (evaluating 'ss.cssRules[0].cssRules[2]')
FAIL Simple CSSOM manipulation of subrules 5 assert_throws_dom: function "() => { ss.cssRules[0].insertRule('% {}'); }" threw object "TypeError: ss.cssRules[0].insertRule is not a function. (In 'ss.cssRules[0].insertRule('% {}')', 'ss.cssRules[0].insertRule' is undefined)" that is not a DOMException SyntaxError: property "code" is equal to undefined, expected 12
FAIL Simple CSSOM manipulation of subrules 6 undefined is not an object (evaluating 'ss.cssRules[0].cssRules[1]')
FAIL Simple CSSOM manipulation of subrules 7 ss.cssRules[0].insertRule is not a function. (In 'ss.cssRules[0].insertRule('@supports selector(&) { & div { font-size: 10px; }}', 1)', 'ss.cssRules[0].insertRule' is undefined)
FAIL Simple CSSOM manipulation of subrules 8 assert_equals: color is changed, new rule is ignored expected ".a {\n  color: olivedrab;\n  & .b { color: green; }\n  & .c { color: blue; }\n}" but got ".a { color: olivedrab; }"
